还在尝试实现查询装备功能

+++ src/main/java/cn/wowtw_backend/model/GearExtraDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class GearExtraDTO {
     private String icon;
     private String description;
-    private Integer quality;
+    private Byte quality;
 }
